requirements.txt
streamlit
google-generativeai
python-dotenv
requests
numpy
Pillow
sounddevice
scipy
